// gen/ts/open_splunk/value.ts
export enum ValueType {
  VALUE_TYPE_UNSPECIFIED = 0,
  VALUE_TYPE_NULL = 1,
  VALUE_TYPE_STRING = 2,
  VALUE_TYPE_SINT64 = 3,
  VALUE_TYPE_UINT64 = 4,
  VALUE_TYPE_DOUBLE = 5,
  VALUE_TYPE_BOOL = 6,
  VALUE_TYPE_BYTES = 7,
  VALUE_TYPE_TIMESTAMP = 8,
  VALUE_TYPE_DURATION = 9,
  VALUE_TYPE_LIST = 10,
  VALUE_TYPE_OBJECT = 11,
  VALUE_TYPE_DECIMAL = 12,
  VALUE_TYPE_MIXED = 13,
  VALUE_TYPE_MISSING = 14,
  UNRECOGNIZED = -1,
}